# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VoidRunner is a distributed task execution platform that allows users to create, manage, and execute code tasks securely in isolated containers.",
        "title": "VoidRunner API",
        "termsOfService": "https://voidrunner.com/terms",
        "contact": {
            "name": "VoidRunner Support",
            "url": "https://github.com/voidrunnerhq/voidrunner",
            "email": "support@voidrunner.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks/{task_id}/executions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts execution of the specified task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Executions"
                ],
                "summary": "Start task execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Execution started successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.TaskExecutionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid task ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Task is already running",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_voidrunnerhq_voidrunner_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "validation_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ValidationError"
                    }
                }
            }
        },
        "github_com_voidrunnerhq_voidrunner_internal_models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "timeout",
                "cancelled"
            ],
            "x-enum-varnames": [
                "ExecutionStatusPending",
                "ExecutionStatusRunning",
                "ExecutionStatusCompleted",
                "ExecutionStatusFailed",
                "ExecutionStatusTimeout",
                "ExecutionStatusCancelled"
            ]
        },
...
        "github_com_voidrunnerhq_voidrunner_internal_models.TaskExecutionResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "execution_time_ms": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "memory_usage_bytes": {
                    "type": "integer"
                },
                "return_code": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_voidrunnerhq_voidrunner_internal_models.ExecutionStatus"
                },
                "stderr": {
                    "type": "string"
                },
                "stdout": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_voidrunnerhq_voidrunner_internal_models.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}